+++ apx_sdk/CHANGELOG.md
@@ -6,12 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Re-exported `Host` from `http_url_whatwg` module.
+- Added `no_proxy` parameter to `FederationAgent`.
+
 ### Changed
 
 - Changed MSRV to 1.80.0.
 - Migrated to `thiserror` 2.0.
 - Replaced `required_components` argument in `parse_http_signature_rfc9421` with a boolean flag.
 - Don't try to extract content digest if request method is not POST, PUT or PATCH.
+- Changed return type of `get_hostname` to `Host<String>`.
 
 ### Security